# Vectorize RandomFlip