--- a/examples/layers/preprocessing/bounding_box/random_flip_demo.py
+++ b/examples/layers/preprocessing/bounding_box/random_flip_demo.py
@@ -1,4 +1,4 @@
-# Copyright 2022 The KerasCV Authors
+# Copyright 2023 The KerasCV Authors
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -22,10 +22,12 @@ from keras_cv.layers import preprocessing
 
 
 def main():
-    dataset = demo_utils.load_voc_dataset(bounding_box_format="xywh")
-    random_flip = preprocessing.RandomFlip(bounding_box_format="xywh")
+    dataset = demo_utils.load_voc_dataset(bounding_box_format="xyxy")
+    random_flip = preprocessing.RandomFlip(
+        mode="horizontal_and_vertical", bounding_box_format="xyxy"
+    )
     result = dataset.map(random_flip, num_parallel_calls=tf.data.AUTOTUNE)
-    demo_utils.visualize_data(result, bounding_box_format="xywh")
+    demo_utils.visualize_data(result, bounding_box_format="xyxy")
 
 
 if __name__ == "__main__":