--- a/keras_cv/layers/preprocessing/random_flip.py
+++ b/keras_cv/layers/preprocessing/random_flip.py
@@ -46,18 +46,19 @@ class RandomFlip(VectorizedBaseImageAugmentationLayer):
         3D (unbatched) or 4D (batched) tensor with shape:
         `(..., height, width, channels)`, in `"channels_last"` format.
 
-    Arguments:
-      mode: String indicating which flip mode to use. Can be `"horizontal"`,
-        `"vertical"`, or `"horizontal_and_vertical"`, defaults to
-        `"horizontal"`. `"horizontal"` is a left-right flip and `"vertical"` is
-        a top-bottom flip.
-      rate: A float that controls the frequency of flipping. 1.0 indicates that
-        images are always flipped. 0.0 indicates no flipping. Defaults to 0.5.
-      seed: Integer. Used to create a random seed.
-      bounding_box_format: The format of bounding boxes of input dataset.
-        Refer to
-        https://github.com/keras-team/keras-cv/blob/master/keras_cv/bounding_box/converters.py
-        for more details on supported bounding box formats.
+    Args:
+        mode: String indicating which flip mode to use. Can be `"horizontal"`,
+            `"vertical"`, or `"horizontal_and_vertical"`, defaults to
+            `"horizontal"`. `"horizontal"` is a left-right flip and
+            `"vertical"` is a top-bottom flip.
+        rate: A float that controls the frequency of flipping. 1.0 indicates
+            that images are always flipped. 0.0 indicates no flipping.
+            Defaults to 0.5.
+        seed: Integer. Used to create a random seed.
+        bounding_box_format: The format of bounding boxes of input dataset.
+            Refer to
+            https://github.com/keras-team/keras-cv/blob/master/keras_cv/bounding_box/converters.py
+            for more details on supported bounding box formats.
     """
 
     def __init__(