# What does this PR do?

## Benchmarks

Vectorized RandomFlip is super fast compared to old one.

||Mode|n=100|n=200|n=500|n=1000|
|-|-|-|-|-|-|
|OldRandomFlip|Eager|0.703s|1.556s|8.011s|29.660s|
|OldRandomFlip|Graph|0.311s|1.191s|7.337s|29.395s|
|RandomFlip|Eager|0.005s|0.006s|0.008s|0.011s|
|RandomFlip|Graph|0.001s|0.002s|0.003s|0.006s|
|RandomFlip|XLA|0.001s|0.002s|0.003s|0.006s|

![comparison](https://user-images.githubusercontent.com/20734616/230883608-5daf69fd-dea9-47c8-b386-61b3e4983aae.png)

OldRandomFlip is too slow, therefore I have set `n` to `[100, 200, 500, 1000]`. OldRandomFlip cannot run on XLA.

## Visualization

This PR updates the working pipeline for visualization.
`mode="horizontal_and_vertical"`

![output](https://user-images.githubusercontent.com/20734616/230883491-632c146b-fa46-4326-b6e4-b5eb9490271f.png)

## Notes

This PR fixes some missing unit tests for RandomFlip.

Vectorized RandomFlip can be run on XLA and provide a speed-up.

## Before submitting
- [ ] This PR fixes a typo or improves the docs (you can dismiss the other checks if that's the case).
- [X] Did you read the [contributor guideline](https://github.com/keras-team/keras-cv/blob/master/.github/CONTRIBUTING.md),
      Pull Request section?
- [ ] Was this discussed/approved via a Github issue? Please add a link
      to it if that's the case.
- [X] Did you write any new necessary tests?
- [ ] If this adds a new model, can you run a few training steps on TPU in Colab to ensure that no XLA incompatible OP are used?

## Who can review?
@LukeWood @ianstenbit