@james77777778 2023-04-11
Hi @LukeWood 

I have upgraded `rate` with a default value of `0.5` to a constructor argument.